+++ src/latrend/methods/__init__.py
@@ -1,6 +1,6 @@
 from .features import lcMethodFeatures
+from .kml import lcMethodKML
 from .lmkm import lcMethodLMKM
 from .random import lcMethodRandom
 
-__all__ = ["lcMethodRandom", "lcMethodLMKM", "lcMethodFeatures"]
-
+__all__ = ["lcMethodRandom", "lcMethodLMKM", "lcMethodFeatures", "lcMethodKML"]